requirements.txt
antlr4-python3-runtime==4.9.1

Flask==2.0.1
Flask-Cors==3.0.10

graphviz==0.17

kafka-python==2.0.2

numpy==1.20.1
openpyxl==3.0.7
pandas==1.2.4

pyan3==1.2.0
pycparser==2.20
qrcode==6.1
requests==2.26.0

xlrd==1.2.0

